style: New slice_new.go file
