+++ internal/pkg/service/buffer/sink/tablesink/storage/repository/slice_repo_test.go
@@ -15,12 +15,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition/column"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition/key"
 	defRepository "github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition/repository"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/compression"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/test"
 	deps "github.com/keboola/keboola-as-code/internal/pkg/service/common/dependencies"
 	serviceError "github.com/keboola/keboola-as-code/internal/pkg/service/common/errors"
@@ -441,43 +438,4 @@ storage/slice/level/target/123/456/my-source/my-sink-1/2000-01-03T02:00:00.000Z/
 }
 >>>>>
 `
-}
-
-func TestNewSlice_InvalidCompressionType(t *testing.T) {
-	t.Parallel()
-
-	// Fixtures
-	now := utctime.MustParse("2000-01-01T19:00:00.000Z").Time()
-	projectID := keboola.ProjectID(123)
-	branchKey := key.BranchKey{ProjectID: projectID, BranchID: 456}
-	sourceKey := key.SourceKey{BranchKey: branchKey, SourceID: "my-source"}
-	sinkKey := key.SinkKey{SourceKey: sourceKey, SinkID: "my-sink-1"}
-	fileKey := storage.FileKey{SinkKey: sinkKey, FileID: storage.FileID{OpenedAt: utctime.From(now)}}
-	cfg := storage.NewConfig()
-	mapping := definition.TableMapping{
-		TableID: keboola.MustParseTableID("in.bucket.table"),
-		Columns: column.Columns{
-			column.Datetime{Name: "datetime"},
-			column.Body{Name: "body"},
-		},
-	}
-	credentials := &keboola.FileUploadCredentials{
-		S3UploadParams: &s3.UploadParams{
-			Credentials: s3.Credentials{
-				Expiration: iso8601.Time{Time: now.Add(time.Hour)},
-			},
-		},
-	}
-
-	// Create file
-	file, err := newFile(fileKey, cfg, mapping, credentials)
-	require.NoError(t, err)
-
-	// Set unsupported compression type
-	file.LocalStorage.Compression.Type = compression.TypeZSTD
-
-	// Assert
-	_, err = newSlice(now, file, "my-volume", 0)
-	require.Error(t, err)
-	assert.Equal(t, `file compression type "zstd" is not supported`, err.Error())
 }
